organize styles into different files
